--- src/components/profiles/perfilGrid.module.scss
@@ -1,163 +1,0 @@
-$background-color: #f4f4f4;
-$card-radius: 15px;
-$shadow-color: rgba(0, 0, 0, 0.1);
-$primary-color: #333;
-$spacing: 30px;
-$max-width: 100%;
-$button-bg-color: #007bff;
-$button-color: white;
-
-.perfil_grid_container {
-  margin: auto;
-  width: 70%;
-}
-
-.perfil_grid {
-  display: grid;
-  grid-auto-flow: column; /* Dispone los elementos en columnas */
-  grid-template-rows: repeat(2, 1fr); /* Máximo de 2 filas por columna */
-  gap: $spacing;
-  width: max-content;
-  padding: $spacing;
-  position: relative;
-}
-
-.perfil_card {
-  background-color: $background-color;
-  border-radius: $card-radius;
-  text-align: center;
-  padding: $spacing;
-  box-shadow: 0 4px 20px $shadow-color;
-  width: 100%;
-  height: 250px;
-  position: relative;
-  overflow: hidden;
-  transition: transform 0.2s ease;
-
-  &:hover {
-    transform: scale(1.05);
-  }
-
-  img {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-  }
-
-  .perfil_name {
-    display: none;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    font-size: 1.6rem;
-    color: $primary-color;
-    background-color: #d3d3d3;
-    padding: 15px 25px;
-    border-radius: 10px;
-    text-align: center;
-    width: auto;
-  }
-
-  &:hover .perfil_name {
-    display: block;
-  }
-}
-
-.add_button {
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background-color: $button-bg-color;
-  color: $button-color;
-  border: none;
-  padding: 15px 25px;
-  font-size: 1.4rem;
-  border-radius: 10px;
-  cursor: pointer;
-  box-shadow: 0 4px 15px rgba(0, 0, 0, 0.15);
-  transition: background-color 0.3s ease;
-
-  &:hover {
-    background-color: darken($button-bg-color, 10%);
-  }
-}
-
-// Estilos para el modal
-.modal {
-  position: fixed;
-  top: 0;
-  left: 0;
-  right: 0;
-  bottom: 0;
-  background: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 1000;
-}
-
-.modal_content {
-  background: white;
-  padding: 30px;
-  border-radius: 15px;
-  width: 400px;
-  display: flex;
-  flex-direction: column;
-  gap: 20px;
-}
-
-.modal_buttons {
-  display: flex;
-  justify-content: space-between;
-  button {
-    padding: 10px 20px;
-    cursor: pointer;
-    border-radius: 5px;
-  }
-
-  button:nth-child(1) {
-    background-color: #ccc;
-  }
-
-  button:nth-child(2) {
-    background-color: $button-bg-color;
-    color: white;
-  }
-
-  button:disabled {
-    background-color: #d3d3d3;
-    cursor: not-allowed;
-  }
-}
-
-// Estilos para la selección de imágenes
-.image_selector {
-  display: flex;
-  flex-direction: column;
-}
-
-.image_list {
-  display: flex;
-  gap: 10px;
-  margin-top: 10px;
-
-  .image_item {
-    width: 64px;
-    height: 64px;
-    object-fit: cover;
-    border-radius: 8px;
-    cursor: pointer;
-    border: 2px solid transparent;
-    transition: transform 0.2s ease, border-color 0.2s ease;
-
-    &:hover {
-      transform: scale(1.1);
-    }
-
-    &.selected {
-      border-color: #007bff;
-    }
-  }
-}
